+++ CHANGES.md
@@ -3,7 +3,7 @@
 
 Changes in 3.0.2
 ----------------
-- Retry once on 0 error AJAX error, maybe aleatory in warehouse.js
+- Retry once on AJAX error 0, maybe a micro Wifi interruption in warehouse.js
 - Better check if #body should be updated in Side.php
 - Fix Warehouse footer: always open menu to modname in Warehouse.php
 - Cache <script> resources loaded in AJAX in warehouse.js